Refactor tone_detect_demo into proper Python module structure
- Create contact module with __init__.py
- Extract LinkStateTracker to link_state.py
- Extract StatusDisplay to display.py
- Extract tone generation to tone_generation.py
- Extract detection logic to detection.py
- Extract audio setup to audio_setup.py
- Move configuration to config.py
- Simplify main demo script to use module imports
- Add proper module docstrings and structure
- Fix import paths for all modules
